kafka-neo4j-pipeline/5-producer-app/requirements.txt
kafka-python==2.0.2
kafka-python==2.0.2
neo4j==5.12.0